requirement.txt
flask
google-cloud-storage
google-cloud-aiplatform
google-cloud-firestore
requests